♻️ ignore SSR for RSC rendering
