+++ src/components/custom-rsc-renderer/rsc-client-renderer.tsx
@@ -1,19 +1,18 @@
 "use client";
 import * as React from "react";
-import * as RSDWSSr from "react-server-dom-webpack/client.edge";
 import * as RSDW from "react-server-dom-webpack/client";
-
-import { getSSRManifest } from "./rsc-manifest";
+import { unstable_postpone as postpone } from "react";
 
 export type RscClientRendererProps = {
   payloadOrPromise: string | Promise<string>;
-  withSSR?: boolean;
 };
 
 export function RscClientRenderer({
-  payloadOrPromise,
-  withSSR = false
+  payloadOrPromise
 }: RscClientRendererProps) {
+  if (typeof window === "undefined") {
+    postpone("This component can only be used on the client");
+  }
   const renderPromise = React.useMemo(() => {
     /**
      * This is to fix a bug that happens sometimes in the SSR phase,
@@ -24,7 +23,7 @@ export function RscClientRenderer({
      * these fields are used internally by `use` and are what's
      * prevent `use` from suspending indefinitely.
      */
-    const pendingPromise = resolveElement(payloadOrPromise, withSSR)
+    const pendingPromise = resolveElement(payloadOrPromise)
       .then((value) => {
         // @ts-expect-error
         if (pendingPromise.status === "pending") {
@@ -46,7 +45,7 @@ export function RscClientRenderer({
     // @ts-expect-error
     pendingPromise.status = "pending";
     return pendingPromise;
-  }, [payloadOrPromise, withSSR]);
+  }, [payloadOrPromise]);
 
   return <RscClientRendererUse promise={renderPromise} />;
 }
@@ -57,31 +56,13 @@ function RscClientRendererUse(props: {
   return React.use(props.promise);
 }
 
-async function resolveElement(
-  payloadOrPromise: string | Promise<string>,
-  ssr = false
-) {
-  console.log("Render payload to JSX");
+async function resolveElement(payloadOrPromise: string | Promise<string>) {
   const payload =
     typeof payloadOrPromise === "string"
       ? payloadOrPromise
       : await payloadOrPromise;
   const rscStream = transformStringToReadableStream(payload);
-  let rscPromise: Promise<React.JSX.Element> | null = null;
-
-  // Render to HTML
-  if (ssr && typeof window === "undefined") {
-    // the SSR manifest contains all the client components that will be SSR'ed
-    // And also how to import them
-    rscPromise = RSDWSSr.createFromReadableStream(rscStream, getSSRManifest());
-  }
-
-  // Hydrate or CSR
-  if (rscPromise === null) {
-    rscPromise = RSDW.createFromReadableStream(rscStream, {});
-  }
-
-  return await rscPromise;
+  return await RSDW.createFromReadableStream(rscStream, {});
 }
 
 export function transformStringToReadableStream(input: string) {
