+++ src/components/custom-rsc-renderer/rsc-manifest.ts
@@ -1,7 +1,4 @@
-import { evaluateClientReferences } from "~/components/cache";
-
 export async function getClientManifest() {
-  await evaluateClientReferences();
   let clientManifest: ClientManifest = {};
 
   // we concatennate all the manifest for all pages
@@ -15,25 +12,4 @@ export async function getClientManifest() {
     }
   }
   return clientManifest;
-}
-export function getSSRManifest() {
-  let rscManifest: RSCManifest = {};
-
-  // we concatennate all the manifest for all pages
-  if (globalThis.__RSC_MANIFEST) {
-    const allManifests = Object.values(globalThis.__RSC_MANIFEST);
-    for (const manifest of allManifests) {
-      rscManifest = {
-        ...rscManifest,
-        ...manifest
-      };
-    }
-  }
-
-  return {
-    ssrManifest: {
-      moduleLoading: rscManifest?.moduleLoading,
-      moduleMap: rscManifest?.ssrModuleMapping
-    }
-  };
 }
